Relax mbarrier + TMA wrapper signatures to Integer
\`mbarrier.{init,arrive.noComplete,arrive.expect_tx,expect_tx,
test_wait,test_wait.parity,try_wait,try_wait.parity}\` count/tx_count/
phase/state params and \`cp.async.bulk.tensor.<N>d\` coord/mask params
all accept \`Integer\` instead of pinning to \`UInt32\` / \`Int32\` / \`UInt16\`.
The wrappers convert at the @asmcall boundary so the underlying asm
constraint and LLVM type are unchanged.

Eliminates ~40 explicit casts (\`Int32(0)\`, \`UInt32(128)\`, \`UInt16(0x3)\`)
at kernel call sites across the Hopper tests, without changing semantics
or compiled output (the conversion folds at compile time when the caller
passes a literal).

+++ src/wrappers/mbarrier.jl
@@ -53,13 +53,13 @@ end
 
 @generated function (::Operation{:mbarrier, (:init, :shared, :b64)})(
         mbar::Core.LLVMPtr{T, AS.Shared},
-        count::UInt32) where T
+        count::Integer) where T
     spec = mbarrier_spec(:init, :addr_count)
     quote
         Base.@inline
         @asmcall($(spec.asm), $(spec.constraints), true, Nothing,
                  Tuple{Core.LLVMPtr{$T, AS.Shared}, UInt32},
-                 mbar, count)
+                 mbar, UInt32(count))
         nothing
     end
 end
@@ -91,39 +91,39 @@ end
 # Does NOT close the phase even if pending hits zero.
 @generated function (::Operation{:mbarrier, (:arrive, :noComplete, :shared, :b64)})(
         mbar::Core.LLVMPtr{T, AS.Shared},
-        count::UInt32) where T
+        count::Integer) where T
     spec = mbarrier_spec(Symbol("arrive.noComplete"), :state_addr_n)
     quote
         Base.@inline
         @asmcall($(spec.asm), $(spec.constraints), true,
                  UInt64, Tuple{Core.LLVMPtr{$T, AS.Shared}, UInt32},
-                 mbar, count)
+                 mbar, UInt32(count))
     end
 end
 
 # (sm_90+) Fused expect+arrive: records expected-tx-bytes and bumps pending by 1.
 @generated function (::Operation{:mbarrier, (:arrive, :expect_tx, :shared, :b64)})(
         mbar::Core.LLVMPtr{T, AS.Shared},
-        tx_count::UInt32) where T
+        tx_count::Integer) where T
     spec = mbarrier_spec(Symbol("arrive.expect_tx"), :state_addr_n)
     quote
         Base.@inline
         @asmcall($(spec.asm), $(spec.constraints), true,
                  UInt64, Tuple{Core.LLVMPtr{$T, AS.Shared}, UInt32},
-                 mbar, tx_count)
+                 mbar, UInt32(tx_count))
     end
 end
 
 # (sm_90+) Standalone form: no arrive, no state output.
 @generated function (::Operation{:mbarrier, (:expect_tx, :shared, :b64)})(
         mbar::Core.LLVMPtr{T, AS.Shared},
-        tx_count::UInt32) where T
+        tx_count::Integer) where T
     spec = mbarrier_spec(:expect_tx, :addr_count)
     quote
         Base.@inline
         @asmcall($(spec.asm), $(spec.constraints), true, Nothing,
                  Tuple{Core.LLVMPtr{$T, AS.Shared}, UInt32},
-                 mbar, tx_count)
+                 mbar, UInt32(tx_count))
         nothing
     end
 end
@@ -131,13 +131,13 @@ end
 # Token form: pass the UInt64 returned by a prior arrive on the same mbar.
 @generated function (::Operation{:mbarrier, (:test_wait, :shared, :b64)})(
         mbar::Core.LLVMPtr{T, AS.Shared},
-        state::UInt64) where T
+        state::Integer) where T
     spec = mbarrier_spec(:test_wait, :pred_addr_state)
     quote
         Base.@inline
         @asmcall($(spec.asm), $(spec.constraints), true,
                  Bool, Tuple{Core.LLVMPtr{$T, AS.Shared}, UInt64},
-                 mbar, state)
+                 mbar, UInt64(state))
     end
 end
 
@@ -145,13 +145,13 @@ end
 # cycle has completed.
 @generated function (::Operation{:mbarrier, (:test_wait, :parity, :shared, :b64)})(
         mbar::Core.LLVMPtr{T, AS.Shared},
-        phase::UInt32) where T
+        phase::Integer) where T
     spec = mbarrier_spec(Symbol("test_wait.parity"), :pred_addr_phase)
     quote
         Base.@inline
         @asmcall($(spec.asm), $(spec.constraints), true,
                  Bool, Tuple{Core.LLVMPtr{$T, AS.Shared}, UInt32},
-                 mbar, phase)
+                 mbar, UInt32(phase))
     end
 end
 
@@ -159,25 +159,25 @@ end
 # of returning false immediately.
 @generated function (::Operation{:mbarrier, (:try_wait, :shared, :b64)})(
         mbar::Core.LLVMPtr{T, AS.Shared},
-        state::UInt64) where T
+        state::Integer) where T
     spec = mbarrier_spec(:try_wait, :pred_addr_state)
     quote
         Base.@inline
         @asmcall($(spec.asm), $(spec.constraints), true,
                  Bool, Tuple{Core.LLVMPtr{$T, AS.Shared}, UInt64},
-                 mbar, state)
+                 mbar, UInt64(state))
     end
 end
 
 @generated function (::Operation{:mbarrier, (:try_wait, :parity, :shared, :b64)})(
         mbar::Core.LLVMPtr{T, AS.Shared},
-        phase::UInt32) where T
+        phase::Integer) where T
     spec = mbarrier_spec(Symbol("try_wait.parity"), :pred_addr_phase)
     quote
         Base.@inline
         @asmcall($(spec.asm), $(spec.constraints), true,
                  Bool, Tuple{Core.LLVMPtr{$T, AS.Shared}, UInt32},
-                 mbar, phase)
+                 mbar, UInt32(phase))
     end
 end
 
@@ -201,14 +201,14 @@ end
 
 @generated function (::Operation{:mbarrier, (:arrive, :expect_tx, Symbol("shared::cluster"), :b64)})(
         mbar::Core.LLVMPtr{T, AS.Shared},
-        tx_count::UInt32) where T
+        tx_count::Integer) where T
     spec = mbarrier_spec(Symbol("arrive.expect_tx"), :sink_addr_n;
                          ss = Symbol("shared::cluster"))
     quote
         Base.@inline
         @asmcall($(spec.asm), $(spec.constraints), true, Nothing,
                  Tuple{Core.LLVMPtr{$T, AS.Shared}, UInt32},
-                 mbar, tx_count)
+                 mbar, UInt32(tx_count))
         nothing
     end
 end
